% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binary.R, R/export_ssn.R
\name{assign_binIDs}
\alias{assign_binIDs}
\title{assign_binIDs
Recursive function to assign binary id to stream segments}
\usage{
assign_binIDs(dt, id, binID, lastbit)

assign_binIDs(dt, id, binID, lastbit)
}
\arguments{
\item{id:}{stream segment}

\item{binID:}{binary ID}

\item{lastbit:}{last char to be added (0 or 1)}

\item{id:}{stream segment}

\item{binID:}{binary ID}

\item{lastbit:}{last char to be added (0 or 1)}
}
\description{
Should be run for all outlets in the network ( = most downstream segments) and fills the binID for all segments

This function is run for all outlets in the network ( = most downstream segments)
and fills the binID for all segments. It is called by \code{calc_binary} in
\code{export_ssn} and should not be called by the user.
}
\keyword{internal}
