% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_sites_approx.R
\name{calc_attributes_sites_approx}
\alias{calc_attributes_sites_approx}
\title{Calcuate attributes of the sites.}
\usage{
calc_attributes_sites_approx(sites_map = "sites", input_attr_name,
  output_attr_name = NULL, stat, round_dig = 2)
}
\arguments{
\item{sites_map}{string giving the name of the sites the attributes shall be
calculated for. "sites" refers to the observation sites.}

\item{input_attr_name}{character vector; input column name in the edges
attribute data table.}

\item{output_attr_name}{character vector (optional); output column name appended
to the site attribute data table. If not provided it is set to
\code{input_attr_name}. Please not that attribute names must not be longer
 than 10 characters.}

\item{stat}{name or character vector giving the statistics to be calulated. If
it is one of min, max, mean or percent the function returns the same value for
the site as for the edge it lies on. Otherwise, it returns the sum of all
catchments upstream of the last junction upstream of the site, and a proportional
value of the edge it lies on based on distRatio.}

\item{round_dig}{integer; number of digits to round results to.}
}
\value{
Nothing. The function appends new columns to the \code{sites_map}
 attribute table
}
\description{
For each site attributes (predictor variables) are derived based on the values
caluclated for the edge the site lies on. This fuction calculates
approximate values for site catchments as described in Peterson & Ver Hoef, 2014:
STARS: An ARCGIS Tollset Used to Calculate the Spatial Infromation Needed to
Fit Spatial Statistical Models to Stream Netwok Data. J. Stat. Softw., 56 (2).
}
\details{
If \code{stat} is one of "min", "max", "mean" or "percent" simply the
value of the edge is assigned. Otherwise, the value is calculated as the sum
of the upstream catchments and the proportional value of the edge the site lies
on (based on distRatio); this is usefull e.g. for counts of dams or waste
water treatment plant.
}
\note{
\code{\link{import_data}}, \code{\link{derive_streams}},
  \code{\link{calc_edges}}, \code{\link{calc_sites}} or
  \code{\link{calc_prediction_sites}} and
  \code{\link{calc_attributes_edges}} must be run before.
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
derive_streams()
#' cj <- check_compl_junctions()
if(cj)
  correct_compl_junctions()
calc_edges()
calc_sites()
execGRASS("r.slope.aspect", flags = c("overwrite","quiet"),
parameters = list(
  elevation = "dem",
  slope = "slope"
  ))
calc_sites_attributes(input_raster = "slope",  stat = "mean", attr_name = "avgSlopeP")
}
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}

