% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_edges.R
\name{calc_catchment_attributes}
\alias{calc_catchment_attributes}
\title{Aggregate attributes for the total catchment of each stream segment.}
\usage{
calc_catchment_attributes(dt, stat, attr_name, round_dig)
}
\arguments{
\item{dt}{data.table of stream topology and attributes per segment.}

\item{stat}{name or character vector giving the statistics to be calulated,
must be one of: min, max, mean, percent.}

\item{attr_name}{name or character vector of column names for the attribute(s)
to be caculated.}

\item{round_dig}{integer; number of digits to round results to. Can be a vector
of different values or just one value for all attributes.}
}
\value{
Nothing. The function changes the values of the columns attr_name in dt.
}
\description{
This function aggregates the attributes of each segment for the total
catchment of each stream segment. It is called within \code{\link{calc_attributes_edges}}
and should not be called by the user.
}

