% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_edges.R
\name{calc_catchment_attributes_rec}
\alias{calc_catchment_attributes_rec}
\title{calc_catchment_attributes_rec
Aggregate attributes for the total catchment of each stream segment.}
\usage{
calc_catchment_attributes_rec(dt, id, stat, attr_name)
}
\arguments{
\item{dt}{data.table of stream topology and attributes per segment.}

\item{id}{integer; 'stream' of outlet segment to start the calculation from.}

\item{stat}{name or character vector giving the statistics to be calulated,
must be one of: min, max, mean, percent.}

\item{attr_name}{name or character vector of column names for the attribute(s)
to be caculated.}
}
\value{
One row data.table with the cummulative number of cells of the total
 catchment of each segment and the values for each attribute and changes the
 values in dt.
}
\description{
Recursive function to calculate the catchment attributes of each stream
segment. It is called by \code{\link{calc_catchment_attributes}} for each
outlet and should not be called by the user.
}
\note{
The values for stats "mean" and "percent" need to be devided by the cummulative
 number of cells of the total catchment in a subsequent step.
}

