% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_edges.R
\name{calc_edges}
\alias{calc_edges}
\title{Calcuate edges for SSN object.}
\usage{
calc_edges(clean = TRUE)
}
\arguments{
\item{clean}{logical; should intermediate layer (rca, reach contributing area)
be removed from GRASS session?}
}
\value{
Nothing. The function produces the following map:
\itemize{
 \item{'edges'} {derived stream segments with computed attributes needed for SSN (vector)}
}
}
\description{
A vector (lines) map 'edges' is derived from 'streams_v' and several
 attributes are assigned.
}
\details{
Steps include:
\itemize{
 \item{Assign unique 'rid' and 'OBJECTID' to each stream segment}
 \item{Find different stream networks in the region and assign 'netID'}
 \item{Calculate segments upstream distance, 'upDist' = flow length
  to the outlet of the network}
 \item{Calculate reach contributing areas (RCA ) per segment, 'rcaArea'
  = subcatchment area of each segment}
 \item{Calculate catchment areas; 'H2OArea' = total catchment area of each
 segment}
 All lengths are rounded to 2 and all areas to 4 decimal places, respectively.
}
}
\note{
\code{\link{setup_grass_environment}}, \code{\link{import_data}} and \code{\link{derive_streams}} must be run before.
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
derive_streams()
#' cj <- check_compl_junctions()
if(cj){
  correct_compl_junctions()
}
calc_edges()
edges <- readVECT('edges', ignore.stderr = TRUE)
head(edges@data)
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(edges, col = 'blue')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}, Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}

