% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_lakes.R
\name{calc_outdist}
\alias{calc_outdist}
\title{Update flow length from the upstream node of each stream segment 
  to the outlet of the network .}
\usage{
calc_outdist(dt, id)
}
\arguments{
\item{dt}{data.table containing the attributes of the stream segments}

\item{id}{integer; 'stream' of the stream segment}
}
\value{
nothing
}
\description{
Recursive function to calculate the flow length from the 
upstream node of the stream segment to the outlet of the network.
It is called by \code{\link{delete_lakes}} for each
outlet and should not be called by the user.
}
\details{
calc_outdist
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}
\keyword{internal}
