% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compl_junctions.R
\name{check_compl_junctions}
\alias{check_compl_junctions}
\title{Check if there are more than two inflows to an outflow.}
\usage{
check_compl_junctions()
}
\value{
TRUE if there are complex junctions.
}
\description{
It is checked, if there are columns named 'prev_str03', 'prev_str04' and
'prev_str05' in the attribute table of streams_v derived with
\href{https://grass.osgeo.org/grass70/manuals/addons/r.stream.order.html}{r.stream.order},
hence, if there are more than two inflows to a junction.
}
\note{
\code{\link{setup_grass_environment}}, \code{\link{import_data}} and
  \code{\link{derive_streams}} must be run before.
}
\examples{
\donttest{
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
derive_streams()
check_compl_junctions()
}
}
\author{
Mira Kattwinkel \email{mira.kattwinkel@gmx.net}
}

