% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ssn.R
\name{check_ssn}
\alias{check_ssn}
\title{Checking SSN object.}
\usage{
check_ssn(path)
}
\arguments{
\item{path}{character; path to .ssn object.}
}
\value{
TRUE or FALSE depending if checks pass.
}
\description{
This function roughly checks
}
\examples{
\donttest{
# Initiate GRASS session
if(.Platform$OS.type == "windows"){
  gisbase = "c:/Program Files/GRASS GIS 7.2.0"
  } else {
  gisbase = "/usr/lib/grass72/"
  }
initGRASS(gisBase = gisbase,
    home = tempdir(),
    override = TRUE)

# Load files into GRASS
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
gmeta()

# Derive streams from DEM
derive_streams(burn = 0, accum_threshold = 700, condition = TRUE, clean = TRUE)

# Check and correct complex junctions (there are no complex juctions in this 
# example date set)
cj <- check_compl_junctions()
if(cj){
  correct_compl_junctions()
}

# Prepare edges
calc_edges()

# Prepare site
calc_sites()

# Write data to SSN Folder
ssn_dir <- file.path(tempdir(), 'nc.ssn')
export_ssn(ssn_dir)

# Check if all files are ok
library(SSN)
check_ssn(ssn_dir)
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
