% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_compl_junctions.R
\name{correct_compl_junctions}
\alias{correct_compl_junctions}
\title{Correct junctions with three inflows.}
\usage{
correct_compl_junctions(clean = TRUE, celltoldig = 2)
}
\arguments{
\item{clean}{logical; should intermediate files be removed from GRASS
session?}

\item{celltoldig}{integer; number of digits the cell size dimensions are
rounded to before it is checked whether they are identical}
}
\value{
Nothing. The function changes features in \itemize{
  \item{'streams_v':} {Updated streams with topology (vector)}
  \item{'streams_r':} {Updated stream raster (new cat) (raster)} } and copies
  the oringial to \itemize{ \item{'streams_v_o':} {Originally derived streams
  with topology (vector)} \item{'streams_r_o':} {Originally derived stream
  raster (raster).} }
}
\description{
At complex junctions (i.e. more than two inflows to an outflow), the outflow
is broken into two segments at 1/4 of the DEM's cellsize downstream of the
start using
\href{https://grass.osgeo.org/grass73/manuals/v.edit.html}{v.edit}(tool =
break). Then, the stream with the smallest angle to the outflow is moved to
this new junction using
\href{https://grass.osgeo.org/grass73/manuals/v.edit.html}{v.edit}(tool =
vertexmove). So far, this function works only for junctions with three
inflows, not more.
}
\note{
\code{\link{setup_grass_environment}}, \code{\link{import_data}} and
  \code{\link{derive_streams}} must be run before.
}
\examples{
\donttest{
initGRASS(gisBase = "/usr/lib/grass72/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
derive_streams()
cj <- check_compl_junctions()
if(cj){
  correct_compl_junctions()
}
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
streams <- readVECT('streams_v', ignore.stderr = TRUE)
streams_o <- readVECT('streams_v_o', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(streams_o, col = 'red')
lines(streams, col = 'blue')
}
}
\author{
Mira Kattwinkel \email{mira.kattwinkel@gmx.net}
}
