% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pred.R
\name{create_pred}
\alias{create_pred}
\title{Create prediction sites for .ssn object}
\usage{
create_pred(path, name, n = 1000, type = c("random", "regular"), ...)
}
\arguments{
\item{path}{character; path to .ssn folder}

\item{name}{character; name of prediction shapefile}

\item{n}{numeric; (approximate) sample size}

\item{type}{character; "random" for completely spatial random;
"regular" for regular (systematically aligned) sampling;}
}
\value{
SpatialPoints-Object with prediction sites
}
\description{
Create prediction sites for .ssn object
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}
\seealso{
\code{\link[sp]{spsample}}
}

