% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_lakes.R
\name{delete_lakes}
\alias{delete_lakes}
\title{Delete lakes from stream network}
\usage{
delete_lakes(lakes, keep = TRUE)
}
\arguments{
\item{lakes}{character string or object; path to lake vector file (ESRI shape), 
name of vector map in the GRASS data base or sp or sf data object.}

\item{keep}{boolean; should the original 'streams_v' be saved? Default is TRUE.}
}
\value{
Nothing. The function updates 'streams_v' and (if keep = TRUE) saves 
the original file to streams_v_prev_lakes.
}
\description{
When the stream network is derived from a dem, the streams 
will just cross lakes or ponds. However, the flow is disconnected
here and the relationship between sampling points upstream and
downstream of a lake is not clear. For instance, chemicals could
be retained and temperature altered in a lake. This function intersects
the stream network with a given vector map of lakes; it deletes the stream
segments in the lake, breaks those the cross its borders and 
assigns a new, updated topology.
}
\note{
The column 'out_dist' (flow length from the upstream node of the 
 segment to the outlet of the network) is updated based on the new segment length.
 In contrast, 'cum_length' is not updated as it is no longer needed and will 
 be deleted in calc_edges. 

#'
}
\author{
Mira Kattwinkel  \email{mira.kattwinkel@gmx.net}
}
