% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_streams.R
\name{derive_streams}
\alias{derive_streams}
\title{Derive stream network from DEM.}
\usage{
derive_streams(burn = 5, at = 700, condition = TRUE, clean = TRUE)
}
\arguments{
\item{burn}{numeric; how many meters should the streams be burned into the DEM?}

\item{at}{integer; accumulation threshold to use (i.e. minimum flow
accumulation value that will initiate a new stream).}

\item{condition}{logical; should the DEM be conditioned using
\href{https://grass.osgeo.org/grass70/manuals/addons/r.hydrodem.html}{r.hydrodem}?}

\item{clean}{logical; should intermediate raster layer of imported
streams ("streams_or") be removed from GRASS session?}
}
\value{
Nothing. The function produces the following maps:
\itemize{
 \item{"streams_r"} {derived streams (raster)}
 \item{"streams_v"} {derived streams with topology (vector)}
 \item{"dirs"} {flow directions (raster)}
 \item{"accums"} {accumulation values (raster)}
} Additionally, "dem" is modified if condition = TURE and / or burn > 0.
}
\description{
Streams are derived from a DEM using
\href{https://grass.osgeo.org/grass70/manuals/r.stream.extract.html}{r.stream.extract}.
If a stream network is available (see \code{\link{import_data}}) and
burn > 0 it will be first burned into DEM.
Stream topolology is derived using
\href{https://grass.osgeo.org/grass70/manuals/addons/r.stream.order.html}{r.stream.order}.
}
\note{
\code{\link{setup_grass_environment}} and \code{\link{import_data}}
  must be run before.
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path, sites = sites_path)
import_data(dem = dem_path, sites = sites_path)
derive_streams()
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
streams <- readVECT('streams_v', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(streams, col = 'blue')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}, Mira Kattwinkel \email{mira.kattwinkel@gmx.net}
}

