% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_edges.R
\name{get_all_raster_values}
\alias{get_all_raster_values}
\title{get_all_raster_values
Returns all unique values in the raster}
\usage{
get_all_raster_values(raster_name)
}
\arguments{
\item{raster_name}{name of the raster map}
}
\value{
a vector of all values in the raster
}
\description{
Returns the number of different values in the input raster.
}
\note{
This function is sensitive to MASKs, i.e. if a MASK is present,
 only the part or the raster is processed within the MASK; 
 \href{https://grass.osgeo.org/grass78/manuals/r.mask.html}{r.mask}.
}
\keyword{internal}
