% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_edges.R
\name{get_n_val_raster}
\alias{get_n_val_raster}
\title{get_n_val_raster
Returns the number of differnt values in the raster.}
\usage{
get_n_val_raster(raster_name)
}
\arguments{
\item{raster_name}{name of the raster map}
}
\value{
The range of values in the raster map.
}
\description{
Returns the number of different values in the input raster.
}
\note{
This function is sensitive to MASKs, i.e. if a MASK is present,
 only the part or the raster is processed within the MASK; 
 \href{https://grass.osgeo.org/grass78/manuals/r.mask.html}{r.mask}.
}
\keyword{internal}
