% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_attributes_edges.R
\name{get_raster_range}
\alias{get_raster_range}
\title{get_raster_range
Returns the number of differnt values in the raster.}
\usage{
get_raster_range(raster_name)
}
\arguments{
\item{raster_name}{name of the raster map}
}
\value{
The range of values in the raster map.
}
\description{
Returns the number of different values in the input raster.
}
\keyword{internal}
