% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_edges.R
\name{get_streams_edges_in_catchment}
\alias{get_streams_edges_in_catchment}
\title{get_streams_edges_in_catchment
Returns the stream values of this and all upstream edges}
\usage{
get_streams_edges_in_catchment(dt, str_id)
}
\arguments{
\item{dt}{data.table containing the attributes of the stream segments}

\item{str_id}{integer giving the stream_id ('stream') of the starting edge}
}
\value{
vector of stream values of all upstream edges and the calling one.
}
\description{
Recursive function to get the stream from one segment upstream.
This function is used internally and is not intended to be called by the user.
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}
\keyword{internal}
