% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grass_environment.R
\name{grass_v.to.db}
\alias{grass_v.to.db}
\title{Update attribute table.}
\usage{
grass_v.to.db(map, option, type = "line", columns, format)
}
\arguments{
\item{map}{character; name of the map where values should be uploaded.}

\item{option}{character; what values should be uploaded}

\item{type}{character; feature type (default = "line")}

\item{columns}{character; name of the column top upload data to}

\item{format}{character; data format of the new columns (if it must be created)}
}
\value{
Nothing. Uses \href{https://grass.osgeo.org/grass78/manuals/v.to.db.html}{v.to.db}
to populate attribute values from vector features.
}
\description{
Wrapper for v.to.db to catch errors arising in some GRASS versions
}
\details{
Since different versions of GRASS handle v.to.db differently (older versions <= 7.4 need
 the column to exists. while newer ones create the column) different implementations are
 necessary
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}
