% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_network.R
\name{restict_network}
\alias{restict_network}
\title{Restict edges to certain stream networks}
\usage{
restict_network(sites, keep = TRUE)
}
\arguments{
\item{sites}{name(s) of sites.}

\item{keep}{boolean; should the original 'edges' be saved as 'edges_o'? Default is TRUE.}
}
\value{
Nothing. The function updates 'edges' and (if keep = TRUE) saves 
the original file to 'edges_o'.
}
\description{
When the stream network is derived from a DEM, the network will cover the whole
extent of the DEM input. However, the obervation sites might be restricted
to a certain area, i.e. to certain networks. This functin deletes edges
that belong to networks (based on their netID) without sites (observation or prediction).
This can help to save computation time before calculating edge attributes.
}
\author{
Mira Kattwinkel  \email{mira.kattwinkel@gmx.net}
}
