% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_network.R
\name{restrict_network}
\alias{restrict_network}
\title{Restrict edges to certain stream networks}
\usage{
restrict_network(sites, keep = TRUE)
}
\arguments{
\item{sites}{name(s) of sites.}

\item{keep}{boolean; should the original 'edges' be saved as 'edges_o'? Default is TRUE.}
}
\value{
Nothing. The function updates 'edges' and (if keep = TRUE) saves 
the original file to 'edges_o'.
}
\description{
When the stream network is derived from a DEM, the network will cover the whole
extent of the DEM input. However, the obervation sites might be restricted
to a certain area, i.e. to certain networks. This functin deletes edges
that belong to networks (based on their netID) without sites (observation or prediction).
This can help to save computation time before calculating edge attributes.
}
\examples{
\donttest{
# Initiate GRASS session
if(.Platform$OS.type == "windows"){
  gisbase = "c:/Program Files/GRASS GIS 7.4.0"
  } else {
  gisbase = "/usr/lib/grass74/"
  }
initGRASS(gisBase = gisbase,
    home = tempdir(),
    override = TRUE)

# Load files into GRASS
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
lakes_path <- system.file("extdata", "nc", "lakes.shp", package = "openSTARS")
setup_grass_environment(dem = dem_path)
import_data(dem = dem_path, sites = sites_path)
gmeta()

# Derive streams from DEM
derive_streams(burn = 0, accum_threshold = 100, condition = TRUE, clean = TRUE)

# Check and correct complex junctions (there are complex juctions in the 
# example date set if the accumulation threshold is low)
cj <- check_compl_junctions()
if(cj){
  correct_compl_junctions()
}

calc_edges()
calc_sites()

# plot
edges <- readVECT('edges', ignore.stderr = TRUE)
edges_o <- readVECT('edges_o', ignore.stderr = TRUE)
sites <- readVECT('sites', ignore.stderr = TRUE)
plot(edges_o, col = "lightblue", lwd = 2)
lines(edges, col = "blue4")
points(sites, pch = 16, col = "red")
legend("topright", col = c("red", "lightblue", "blue4"), lty = c(NA, 1,1), 
lwd = c(NA,2,1), pch = c(16,NA,NA),
legend = c("sites", "edges orig", "edges restricted"))
}
}
\author{
Mira Kattwinkel  \email{mira.kattwinkel@gmx.net}
}
