% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSTARS.R
\docType{package}
\name{webchem}
\alias{webchem}
\alias{webchem-package}
\title{openSTARS: An open source implementation of the STARS toolbox.}
\description{
GRASS GIS 7.0 (or greater) with installed addons r.stream.basins, r.stream.distance and r.stream.order
is needed.
}

