% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ssn.R
\name{calc_binary}
\alias{calc_binary}
\title{Calculate binary IDs for stream networks.}
\usage{
calc_binary()
}
\value{
A list with one slot for each network id containing a data frame
with 'rid' and 'binaryID' for each segment belonging to this network.
}
\description{
Calculate binary IDs for each stream network built up by '0' and '1'.
This function is called by \code{\link{export_ssn}} and there is no need for it
be called by the users.
}
\note{
\code{\link{import_data}}, \code{\link{derive_streams}},
  \code{\link{calc_edges}} and \code{\link{calc_sites}} must be run before.
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}; Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}

 @export
}

