% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binary.R
\name{calc_binary}
\alias{calc_binary}
\title{Calculate binary IDs for each stream network}
\usage{
calc_binary()
}
\value{
A list of data.frames (for each network one), with rid and the binary code for each segment.
}
\description{
Calculate binary IDs for each stream network
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
calc_edges()
calc_sites()
binaries <- calc_binary()
head(binaries[[1]])
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

