% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_edges.R
\name{calc_edges}
\alias{calc_edges}
\title{Calculate edges for SSN.}
\usage{
calc_edges(clean = TRUE)
}
\arguments{
\item{clean}{logical; Should intermediate layer be removed from GRASS session?}
}
\value{
Nothing. The function produces the following maps:
\itemize{
 \item{"edges"}{derived stream with computed attributes needed for SSN (vector)}
 \item{"stream_topo"}{stream topology measures (vector)}
}
}
\description{
Calcuate edges for SSN object.

Steps include:
\itemize{
 \item{Assigning unique "rid" and "OBJECTID" to each stream segment}
 \item{Finding different stream networks in the region and assigning "netID"}
 \item{Calculation of stream topology, "topo_dim"}
 \item{Calculation of segments upstream distance ("upDist")}
 \item{Calculation of river contributing areas (RCA) per segment, "rcaAreaKm2"}
 \item{Calculation of watershed areas, "H2OAreaKm2"}
}
}
\note{
\code{\link{import_data}} and \code{\link{derive_streams}} must be run before.
Intermediate layers cleaned are:
\itemize{
 \item{"netID_v"}{network IDs (vector)}
 \item{"netID"}{network IDs (raster)}
 \item{"rca"}{RCA per segement (raster)}
 \item{"streams_r"}{stream network (raster)}
}
}
\examples{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir())
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
calc_edges()
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
edges <- readVECT('edges_tmp', ignore.stderr = TRUE)
head(edges@data)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(edges, col = 'blue')
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

