% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sites.R
\name{calc_sites}
\alias{calc_sites}
\title{Calculate sites for SSN.}
\usage{
calc_sites()
}
\description{
Calcuate sites for SSN object.

Steps include:
\itemize{
 \item{Snap points to derived network. 'dist' give the distance to streams.}
 \item{Assign unique 'pid' and 'locID'.}
 \item{Set underlying 'rid' and 'netID' from stream network.}
 \item{Calculate upstream distance for each point ('upDist')}
 \item{Calculate watershed for each point, 'H2Oarea'}
 \item{Calculate watershed predictors for each point from raster files [currently not implemented]}
}
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
calc_edges()
calc_sites()
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
edges <- readVECT('edges', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(edges, col = 'blue')
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

