% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{check_projection}
\alias{check_projection}
\title{Compare projection raster data to the one of the current GRASS location.}
\usage{
check_projection(path)
}
\arguments{
\item{path}{character string vector; path raster data file(s)}
}
\value{
Nothing.
}
\description{
Check if the procection of raster files matches the one of the 
current region
}
\details{
Prints out a table of the PROJ.4 elements of the projection
  information of the current GRASS location and of the raster file(s) as
  well as one columns for each comparison indicating the differences. 
  Based on this information it can be decided if the data can be read
  into GRASS (\code{\link{import_data}}) without prior processing, i.e.
  if all raster data are of the same projection.
}
