% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ssn.R
\name{check_ssn}
\alias{check_ssn}
\title{Checking SSN object.}
\usage{
check_ssn(path)
}
\arguments{
\item{path}{character; path to .ssn object.}
}
\value{
TRUE or FALSE depending if checks pass.
}
\description{
This function roughly checks
}
\examples{
\donttest{
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
calc_edges()
calc_sites()
binaries <- calc_binary()
ssn_dir <- file.path(tempdir(), 'nc.ssn')
export_ssn(ssn_dir, binary = binaries)
check_ssn(ssn_dir)
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

