% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_streams.R
\name{derive_streams}
\alias{derive_streams}
\title{Derive stream network from DEM.}
\usage{
derive_streams(burn = 5, at = 700, clean = TRUE)
}
\arguments{
\item{burn}{numeric; How many meters should the streams be burned in DEM.}

\item{clean}{logical; Should intermediate layer be removed from GRASS session?}

\item{accumulation}{numeric; accumulation threshold to use.}
}
\value{
Nothing. The function produces the following maps:
\itemize{
 \item{"streams_r"}{derived streams (raster)}
 \item{"streams_v"}{derived streams (vector)}
 \item{"dirs"}{flow directions (raster)}
 \item{"accums"}{accumulation values (raster)}
}
}
\description{
Streams are derived from a DEM using
\href{https://grass.osgeo.org/grass70/manuals/r.stream.extract.html}{r.stream.extract}.
If a stream network is available (see \code{\link{import_data}}) and
burn > 0 it will be first burned into DEM.
}
\note{
\code{\link{import_data}} must be run before.
Intermediate layers cleaned are:
\itemize{
 \item{"streams_or"}{raster of imported streams (raster)}
 \item{"streams_vr"}{uncleaned vector of derived network (vector)}
}
}
\examples{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir())
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites_o', ignore.stderr = TRUE)
streams <- readVECT('streams_v', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
points(sites, pch = 4)
lines(streams, col = 'blue')
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

