% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_snn.R
\name{export_ssn}
\alias{export_ssn}
\title{Export .ssnb object}
\usage{
export_ssn(path, binary)
}
\arguments{
\item{path}{character; path to write .ssn object to.}

\item{binary}{object of type \code{binary}, as created by \code{\link{calc_binary}}.}
}
\value{
Nothing. Files are written to the specified folder
}
\description{
This function takes the calculated sites, edges and binary IDs and exports them
to a folder which then can be read using the SSN package.
}
\examples{
\donttest{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
derive_streams()
calc_edges()
calc_sites()
binaries <- calc_binary()
ssn_dir <- file.path(tempdir(), 'nc.ssn')
export_ssn(ssn_dir, binary = binaries)
list.files(ssn_dir)
}
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

