% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data into GRASS.}
\usage{
import_data(dem, sites, streams = NULL, ...)
}
\arguments{
\item{dem}{character; path to DEM.}

\item{sites}{character; path to sites.}

\item{streams}{character, (optional); path to network shape.
If available it can be to burn into DEM}

\item{...}{other paths to raster data to import as predictors
(currently not implemented)}
}
\value{
Nothing, the data is loaded into the GRASS session.
The DEM is stored as raster 'dem' and sites as vector 'sites_o' within GRASS.
}
\description{
This function load dem and sites data (both required) into the GRASS session.
}
\note{
A GRASS session must be initiated before, see  \code{\link[rgrass7]{initGRASS}}.
}
\examples{
library(rgrass7)
initGRASS(gisBase = "/usr/lib/grass70/",
  home = tempdir(),
  override = TRUE)
gmeta()
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)

dem <- readRAST('dem')
plot(dem)
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

