% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data into GRASS.}
\usage{
import_data(dem, sites, streams = NULL, ...)
}
\arguments{
\item{dem}{character; path to DEM.}

\item{sites}{character; path to sites.}

\item{...}{other paths to raster data to import as predictors
(currently not implemented)}

\item{stream}{character, (optional); path to network shape.
If available it can be to burn into DEM}
}
\value{
Nothing, the data is loaded into the GRASS session.
}
\description{
Import data into GRASS.
}
\note{
A GRASS session must be initiated before, see  \code{\link[rgrass7]{initGRASS}}.
}
\author{
Eduard Szoecs, \email{eduardszoecs@gmail.com}
}

