% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_vector_data}
\alias{import_vector_data}
\title{Generic function to import vector data of various formats into the GRASS environment.}
\usage{
import_vector_data(data, name, proj_dem)
}
\arguments{
\item{data}{character string or object; path to data vector file (shape) 
or sp or sf data object.}

\item{name}{string giving the base name of the vector data to use for the import}

\item{proj_dem}{projection of the dem as project4string}
}
\value{
Nothing.
}
\description{
Generic function to import vector data of various formats into the GRASS environment.
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}
\keyword{internal}
