% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSTARS.R
\name{openSTARS_data}
\alias{openSTARS_data}
\title{Datasets shipped with openSTARS}
\description{
Datasets shipped with openSTARS
}
\section{elev_ned30m.tif}{

South-West Wake county National Elevation Data 30m.
The data has been taken from the GRASS GIS North Carolina data set.
Source \url{https://grass.osgeo.org/download/sample-data/}.
}

\section{lakes.shp}{

Artificial lakes (not at topologically correct locations).
}

\section{sites_nc.shp}{

Arbitrary sites along rivers in North Carolina.
}

