% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_grass_environment.R
\name{setup_grass_environment}
\alias{setup_grass_environment}
\title{Setup 'GRASS' environment.}
\usage{
setup_grass_environment(dem, epsg = NULL, sites = NULL)
}
\arguments{
\item{dem}{character; path to DEM.}

\item{epsg}{integer; EPSG code for the projection to use. If not given (default)
the infromation is taken from the dem}

\item{sites}{(deprecated); not used any more. Only inlcuded for compatibility with previous version.}
}
\value{
Nothing, the 'GRASS' mapset is set to PERMANENT, the projection and
the extent of the current location is set to the one of the dem.
}
\description{
This function sets the 'GRASS' mapset to PERMANENT and sets its projection and extension.
}
\note{
A 'GRASS' session must be initiated before, see \code{\link[rgrass7]{initGRASS}}.
}
\examples{
\donttest{
# Initiate GRASS session
if(.Platform$OS.type == "windows"){
  gisbase = "c:/Program Files/GRASS GIS 7.4.0"
  } else {
  gisbase = "/usr/lib/grass74/"
  }
initGRASS(gisBase = gisbase,
    home = tempdir(),
    override = TRUE)

# Load files into GRASS
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
setup_grass_environment(dem = dem_path)
gmeta()
}
}
\author{
Mira Kattwinkel, \email{mira.kattwinkel@gmx.net}
}
