% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSTARS.R
\docType{package}
\name{webchem}
\alias{webchem}
\alias{webchem-package}
\title{openSTARS: An open source implementation of the STARS toolbox.}
\description{
GRASS GIS 7.0 (or greater) with installed addons r.stream.basins, r.stream.distance and r.stream.order
is needed.
}
\examples{
\donttest{
library(openSTARS)
# Initiate GRASS Sesion
initGRASS(gisBase = "/usr/lib/grass70/",
    home = tempdir(),
    override = TRUE)

# Load files into GRASS
dem_path <- system.file("extdata", "nc", "elev_ned_30m.tif", package = "openSTARS")
sites_path <- system.file("extdata", "nc", "sites_nc.shp", package = "openSTARS")
import_data(dem = dem_path, sites = sites_path)
gmeta()

# Derive streams from DEM
derive_streams()

# Prepare edges
calc_edges()

# Prepare site
calc_sites()

# Calculate binary IDs
binaries <- calc_binary()

# Plot data
dem <- readRAST('dem', ignore.stderr = TRUE)
sites <- readVECT('sites', ignore.stderr = TRUE)
edges <- readVECT('edges', ignore.stderr = TRUE)
plot(dem, col = terrain.colors(20))
lines(edges, col = 'blue')
cols <- colorRampPalette(c("blue", 'red'))(length(sites$value))[rank(sites$value)]
points(sites, pch = 16, col = cols)

# Write SSN Folder
ssn_dir <- file.path(tempdir(), 'nc.ssn')
export_ssn(ssn_dir, binary = binaries)

# Load into SSN-package
library(SSN)
ssn_obj <- importSSN(ssn_dir, o.write = TRUE)
plot(ssn_obj, 'value')
}
}

