% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openSTARS.R
\docType{package}
\name{webchem}
\alias{webchem}
\alias{webchem-package}
\title{openSTARS: An open source implementation of the STARS toolbox.}
\description{
openSTARS: An open source implementation of the STARS toolbox.
}

